Update patch set 1
Patch Set 1: Code-Review+1

I'm more or less positive to this as a method of providing an internal interface for ODL functions to provide/consume toward the Neutron service.  It may make sense to break it up, but if it is internally consumed that would seem less urgent than establishing an interface to use in Lithium.

Patch-set: 1
Reviewer: Gerrit User 503 <503@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1