Update patch set 1
Patch Set 1:

(2 comments)

With my streaming work I don't think we need the EntriesPackagingActor - it will basically be a pass-thru for AppendEntries.

With these changes, the heartbeat message will only serve to prevent election time outs. Currently it calls sendAppendEntries which evaluates what needs to be sent to the follower each time. If installing a snapshot, it sends the next chunk. Also when consensus is reached on AppendEntriesReply and the leader commits the log index and applies the state locally, the next sendAppendEntries heartbeat sends the new commitIndex to the follower which signals it to apply its state locally. Without the heartbeat, these would not occur until the next transaction commit and replication which would be an arbitrary time period. So these changes may have side-effects which may not be desirable.

Patch-set: 1