Update patch set 1
Patch Set 1:

I assume the purpose of moving the heartbeat out of the raft/shard actor is to prevent spurious leader re-election due to the processing of a message taking too long and blocking the heartbeat message.  How about we have the HeartbeatActor send a simple HeartBeat message with no data? The leader code would still do sendAppendEntries on a timer (at some fixed period say 1 sec) as it does now although we could add some smarts to not send an empty message if nothing changed for a follower. The leader would not have to communicate with the HeartbeatActor. We could then increase the default heartbeat time out to like 5 sec.

Patch-set: 1