Update patch set 1
Patch Set 1:

(6 comments)

This patch needs more work. It breaks replication in 2 ways (a) It does ensure via the heartbeat mechanism that an entry is replicated - it relies solely on the replicate message (b) It does not forward AppendEntriesReply to the RaftActor so the RaftActor would never know about the receipt of an AppendEntry.

Patch-set: 1