Update patch set 1
Patch Set 1:

(1 comment)

Tom, I saw your point about the packaging actor not being required in the new code.

So I see that we do process AppendEntriesReply because the Heartbeat actor uses the raft actor as the sender so that is not an issue. Now for the second issue I had of sending replication messages to a lagging follower - a nice optimized solution maybe that when we get an appendentries reply for a given follower we check in the leader if more data needs to be sent to the follower and we send it right away instead of waiting for a timeout. This way we do not need to schedule a heartbeat in the raftactor itself.

Patch-set: 1