Update patch set 3
Patch Set 3:

Yeah - I thought about that. The reason I didn't do it is that you'd lose the desired line # in the log output (assuming you specify the line # in the appender format in the pax logging cfg) as the logger takes the line # of the first frame in the calling stack so you'd always see the log adapter method line #. I've done similar in the past and later regretted it.

Patch-set: 3
Label: Verified=0