Update patch set 3
Patch Set 3:

My understanding is that the LoggingAdapter itself is a wrapper so line number preservation may not be maintained even today. Ultimately the log messages are written using an actor the LoggingAdapter simply makes it easy to log without sending messages to that logging actor and also does things like log level checking and interpolation.

Patch-set: 3