Update patch set 3
Patch Set 3:

(1 comment)

The LoggingAdapter does capture line # but it appears to be from the logging actor so it's always the same - just noticed that. That's not good. Line #'s are very useful. Why do we use akka logging and not slf4j (or whatever we use elsewhere)? Also I tried to enable debug for Shard in the pax logging file but it didn't work. I'm assuming you have to enable it in the akka conf file. If we can't enable debug in the pax logging and have to do it in the akka conf (and restart karaf) then I don't think that's desirable.

Patch-set: 3