Update patch set 1
Patch Set 1:

The scheduler API asks for actorReference, not Actorselection object, we can't use actorselection directly in the API. That is why we need to resolve it.

now, after your streaming code addition, Moiz and I were talking, if we need separate heartbeatactor or not? Can leader itself do the job, like it used to

Patch-set: 1