Update patch set 1
Patch Set 1:

(2 comments)

I looked at the MetricRegistry code. It does appear to be thread safe. getOrAdd accesses 'metrics' which is a ConcurrentMap. The Timer code also appears to the thread safe as it uses concurrent classes and LongAdders. 

The Timer class contains Meter, Reservoir, and Histogram instances which in  turn allocate memory. Also the snapshot allocates a long[]. So added GC might be significant - we might want to sample every nth commit.

Patch-set: 1