Refactor DataStoreContext
It was getting difficult to maintain because we had the same field defined
in the Builder and the DatastoreContext and we were passing the fields along
using the DatastoreContext constructor. This meant that every new setting that
we added required multiple changes to be made to DatastoreContext.

Refactored for the Builder to create the DatastoreContext upfront and directly set
it's fields.

Change-Id: I8060c4ea3dfd02a5cac20e28fb5be2f26b1e1dbd
Signed-off-by: Moiz Raja <moraja@cisco.com>