Update patch set 7
Patch Set 7:

I am not able to reproduce this even after many runs.

One logic i can think of is that in RaftActor#onRecoveryCompletedMessage(), we call onRecoveryComplete() before initializeBehavior().

It could have happened that in the test, the wait for recovery went through and the leader was assigned. But the initializeBehaviour got called later.

Patch-set: 7
Label: Verified=0