Update patch set 8
Patch Set 8:

I have made a change to call initializeBehaviour() to be called before onRecoveryComplete() in RaftActor

I dont think it should affect the recovery complete  in Shard, if the behaviour is set to Follower (which was null initially). 

None of the tests failed and a quick test also did not give any issues.

Hopefully the tests should not break.

Btw, how are you running the tests, is there a script?

Patch-set: 8
Label: Verified=0