Update patch set 8
Patch Set 8:

It failed again on the 92nd run. You had moved the call to initializeBehavior in onRecoveryCompletedMessage() but the test uses a mock DataPersistenceProvider with persistence disabled so it executes the  else clause in RaftActor#handleRecover. The initializeBehavior class needs to be moved here as well.

I'm running the tests with that change - will report the results.

Patch-set: 8