Update patch set 3
Patch Set 3:

(1 comment)

Sending an AppendEntries on AppendEntries creates a situation where we may send a duplicate AppendEntries when the heartbeat triggers. It would be good to avoid that situation by checking when was the last time the AppendEntriesReply was received and if it was too close to the heartbeat time we can skip sending an AppendEntries to that particular follower. Thoughts?

Patch-set: 3
Label: Verified=0