Update patch set 3
Patch Set 3:

(1 comment)

I see your point Moiz - we will end up sending duplicates. That also happens when replications happen quickly.  That's not really a bad thing as follower handles it. Downside is a little more memory used in leader to construct the AppendEntries log list instead of passing emptyList. Also more to serialize over the wire. 

So in handleAppendEntriesReply, check if the current time is close to the next heartbeat. Do we know when the next heartbeat will go off? How close - within half the heartbeat?

Patch-set: 3