Update patch set 8
Patch Set 8:

I am not suggesting a sleep. That was just to confirm if that's indeed the issue.

Regarding the solution, rather than juggling with setting the behaviour, all we want to know is when the initialization is done. 
For that i guess we can have another latch (similar to recovery complete) in the MockRaftActor and introduce a method, waitForInitializeBehaviorComplete(). Since initializeBehavior() is only called for recovery, i believe this should not cause an issue

I have made the changes, ran it 100 times and it didnt fail for me.

Patch-set: 8