Update patch set 5
Patch Set 5:

a qq for my understanding: till now we were using akka's event mechanism to log both the akka's own events and our actor logging as well.
If we move our logging to slf4j, wont this cause a mismatch/mistiming between the logs of akka and ours. 

Should we change the 'loggers' param in akka.conf to point to non-event based slf4j (it atall we can)?

Patch-set: 5
Label: Code-Review=0