Refactor FollowerTest
Removed the Within blocks in tests.

Changed all the tests to use the leaderActor and followerActor class
members for consistency. The actors are killed on tearDown by the TestActorFactory.
The creation of a JavaTestKit block is no longer necessary and was removed from each test.

Added some more assertions in some tests.

Change-Id: Iee793c2b1129b5826dbd3c267c404aae30418fdb
Signed-off-by: tpantelis <tpanteli@brocade.com>