Update patch set 1
Patch Set 1:

Or maybe not. The only thing with your patch is that on replication it still won't send AppendEntries. I think we only want check timeSinceLastActivity if it's initiated by a scheduled heartbeat. So I think we want to go with my patch. What do you guys think?

I can change the unit test.

Patch-set: 1