Bug-2692:Real snapshots should use the replicatedToAllIndex for clearing in-mem log
Similar to fake snapshots, real snapshots should also be using the replicatedToAllIndex.

replicatedToAllIndex is part of the replicatedLog and fakesnapshot helper method is also moved to replicatedlog.

On CaptureSnapshotReply, we clear the log based on replicatedToAllIndex, acquired prior to the snapshot capture.

Many of the existing tests seem to test out this changes and some were enhanced.

Change-Id: Ifb3ca19691f20735e3f84b968a7adf01398c20e0
Signed-off-by: Kamal Rameshan <kramesha@cisco.com>