Update patch set 2
Patch Set 2:

Moiz - it looks like it might be hairy to try to split this up. The tests I modified needed changes due to failures, at the same time I refactored some a bit, e.g. changed ExpectMsg with match/hint to a single line expectMsgClass - the former is harder to read.

How about if I comment each section explaining whether it was a refactor or an actual test change?

I'll try to be better about this in future. When I see code around what I'm changing that can be optimized or made easier, I tend to do it w/o thinking about reviews.

Patch-set: 2