Update patch set 2
Patch Set 2:

> With this setting what happens if you were to restart a node.
 > Especially the first seed-node. Does that work correctly?

Yes. I confirmed with the latest controller code + this change that for all the manual test cases below that a) all Shards eventually elect a new Leader when needed (the rest are Followers) and b) restarted controllers are able to rejoin the cluster within the auto-down time window without any issues.
 
Tests:
1) Take down seed node #1 --> restart seed node #1
2) Take down seed nodes #2 & #3 --> restart seed nodes #2 & #3

Patch-set: 2