Update patch set 4
Patch Set 4:

I had to move the fake-snapshot method out of the replicated log to the abstractraftbehaviour. 
The logic inside this method does not seem to go well with the replicated log and tinkering with the logic would have made the method useless.
I have named the method performSnapshotWithoutCommit so that that the call in Leader and Follower is self-explanatory.

Patch-set: 4
Label: Verified=0