Update patch set 12
Patch Set 12:

That crossed my mind but then you'd have to specify the custom tag in the appender in the pax logging cfg to see it (I believe). Plus wouldn't each log statement have to inject logName into MDC (I think it's TLS)? Or maybe every method would do it at the beginning.

IMO, it's not a big deal to prepend the logName in each log statement. We're not adding log statements a lot. And if we miss one in future, not a big deal.

Patch-set: 12
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0