Update patch set 12
Patch Set 12:

Looking at http://logback.qos.ch/manual/mdc.html, to be safe you'd have to do remove in a try/finally block. This could be done in onReceiveCommand for Shard and RaftActor and handleMessage in each behavior. Still we'd have to specify it in the pax logging file. Maybe a OLD could define a generic logging variable  that other code could use.

Patch-set: 12