Update patch set 5
Patch Set 5:

I haven't looked at the TestDriver in any detail nor have I ever run it. I'm assuming this code was added for initial prototyping which is fine but is it really needed anymore now that we have extensive unit tests and Moiz's dummy data store, et al? It may get to be a burden trying to maintain this code going forward and trying not to break it. I can see this code easily getting way out of date as it's not part of regression. 

If there is useful code in the TestDriver perhaps it should be converted to unit tests or integrated with Moiz's dummy data store.

Patch-set: 5