Update patch set 1
Patch Set 1:

The crypto implementation looks fine to me. Note that the private key is stored in a file that is likely to be accessible to any attacker who is able to access the config XML. Therefore this patch really only adds "security through obscurity", as an attacker could read the private key, then use it to decrypt the password stored in the config XML. Is it possible to store the passwords as salted cryptographic hashes? This would be much stronger from a security perspective, but relies on devices allowing you to authenticate by providing a hashed password rather than a plaintext password. If that is not possible, then I think this patch is a reasonable compromise.

Patch-set: 1
Reviewer: Gerrit User 3339 <3339@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0