Update patch set 6
Patch Set 6: Code-Review-1

According to the published neutron API, 409 is not a valid code to return for the error condition that a UUID already exists.

Second, this change should not be made to these files as they are going away.  Any changes should be made to the neutron repository

Patch-set: 6
Reviewer: Gerrit User 590 <590@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1