Dynamically update DatastoreContext properties
Added a ConfigurationListener to receive ConfigurationEvents from
the Config Admin when a setting in the
org.opendaylight.controller.cluster.datastore.cfg file is modified.

Modified DistributedDataStore so it can be restarted with a new
DatastoreContext. This involves closing the current ActorContext, which
kills the ShardManager and Shard actors, and then creating a new
ActorContext which spins up a new ShardManager actor. In addition we
need to re-register previous DataChangeListeners so the
DistributedDataStore now keeps track of the listener regs and captures,
closes the previous registrations and re-registers them after the new
ActorContext is created.

Change-Id: Ie608f61da36ac58a806208925a3c4277968c2f5b
Signed-off-by: tpantelis <tpanteli@brocade.com>