Dynamically update DatastoreContext properties
Added a ConfigurationListener to the DatastoreContextConfigAdminOverlay
to receive ConfigurationEvents from the Config Admin when the
org.opendaylight.controller.cluster.datastore.cfg file is modified.

The DistributedDataStore registers as a listener to the
DatastoreContextConfigAdminOverlay to get updated when the
DatastoreContext instance changes. The DistributedDataStore sets the
new DatastoreContext instance in the ActorContext which updates its
cached instance and properties and sends the DistributedDataStore
instance as a message to the ShardManager. The ShardManager simply
sends it out to all the Shards which update their cached DatastoreContext
state.

The Shard also sets the new ConfigParams instance in the base RaftActor
which updates its RaftActorContext. There was one place in
FollowerLogInformationImpl where it had a local cache of a property
obtained from the ConfigParams so I changed the ctor to pass the
RaftActorContext instead and obtain the property each time from the
ConfigParams.

A slight downside to this is that we'll have to be cognizant of not caching
DatastoreContext/ConfigParams properties and assuming they're static. Or
if we do cache we need to handle updates.

I toyed around with trying to restart the DistributedDataStore with a
new DatastoreContext and ActorContext and killing the previous
ShardManager and Shard actors and creating new ones. However recreating
the Shards without being disruptive to clients is tricky and risky, eg handling
infight transactions, and I didn't see a clean way to do it. So I went
with the simpler approach of just pushing an updated DatastoreContext to the actors.

Change-Id: Ie608f61da36ac58a806208925a3c4277968c2f5b
Signed-off-by: tpantelis <tpanteli@brocade.com>