Update patch set 11
Patch Set 11:

I am late to this review so I may not have all the background of what may have already discussed before this patchset.

Thinking about how this would be utilized by the clustering system (RAFT) there appears to be enough functionality with this patchset to allow APPs to define the DOM subtree "logical" sharding of their data to control transactions (producers) and listeners (consumers). And with some additional changes to the clustering (RAFT) sharding we should be able to map it to fine-grain RAFT shards to get the matching "physical" data distribution.

APPs that want to fine-tune their clustered performance (and therefore be more cluster-aware by choice) or for various functional reasons (e.g. thinking of per-shard Inventory nodes to represent device ownership, etc.) could do so with this alternative DOMService and future RAFT changed for fine-grained sharding. A single DOMDataTreeProducer a top-level module (e.g. Inventory) with current module-level RAFT Sharding appears to provide the same default behavior we have today, which is great, that we can build on.

I'd be happy to hear the vision for how this will integrate with RAFT sharding. 


Note: For the extreme per-shard Inventory Node case with potentially 1000s of node shards I believe we will have to redefine the level of shard heartbeating.

Patch-set: 11
Reviewer: Gerrit User 2508 <2508@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0