Update patch set 5
Patch Set 5:

I forgot to mention in the commit message some details about DatastoreContextIntrospector. It uses the yang-generated types, NonZeroUint32Type, to validate the property values via reflection. So it will try to construct say a NonZeroUint32Type instance which validates in the constructor and throws IllegalArgEx.

Also if you prepend the data store type ("config" or "operational") to the property name in the file, it will apply it only to that data store type. Without a data store type prefix, it will apply to both. This makes it easier than setting via the config XML file where you have to do it in 2 places. I think persistent is about the only one you'd want to set per data store.

Patch-set: 5
Label: Verified=0