Create odl-nsf-service feature, which excludes neutron feature
Deprecated neutron interfaces in controller, as they are now moved to the neutron repo.

With this change, odl-nsf-service provides an upgrade path, so that projects can migrate to the new neutron
feature w/out changing the other features associated with the soon to be retired odl-nsf-all.

The new neutron features are available via: odl-neutron-service

Change-Id: Ifabf60e62d1f8c512165a8d32d25093e13ff9107
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>