Add RaftActorIntegrationTest
Added RaftActorIntegrationTest with tests for various replication and
snapshotting scenarios. These differ from the RaftActorTest and behavior
tests in that it uses real leaders and followers with real RaftActors
end-to-end.

For the tests, I wanted to use the InMemoryJournal and InMemorySnapshot
implementations in sal-distibuted-datastore, which are close to real, and
I didn't want to duplicate them in sal-akka-raft. So I moved them to
sal-akka-raft and changed the tests in sal-distibuted-datastore to use
them, hence most of the other changes in this patch. sal-akka-raft already
creates a test-jar so I just needed to add it as a dependency in
sal-distibuted-datastore. As a result of this, I removed the
MockAkkaJournal and MockSnapsotStore classes.

I also modified or added logging in some other raft classes that I
thought was useful. I found it useful to output the updated
FollowerLogInformation in AbstractLeader#handleAppendEntriesReply but
only if it changed to avoid unnecessary output for heartbeats. So I changed
FollowerLogInformationImpl to return a boolean from a couple setter
methods if the new index differs.

Change-Id: Ie824bda1f66d1cdb4ce62044d0ba802f4f651496
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>