Add RaftActorIntegrationTest
Added RaftActorIntegrationTest with tests for various replication and
snapshotting scenarios. These differ from the RaftActorTest and behavior
tests in that it uses real leaders and followers with real RaftActors
end-to-end.

For the tests, I wanted to use the InMemoryJournal and InMemorySnapshot
implementations in sal-distibuted-datastore, which are close to real, and
I didn't want to duplicate them in sal-akka-raft. So I moved them to
sal-clustering-commons, hence most of the other changes in this patch.

Change-Id: Ie824bda1f66d1cdb4ce62044d0ba802f4f651496
Signed-off-by: tpantelis <tpanteli@brocade.com>o