Update patch set 1
Patch Set 1:

The test failure is orthogonal to this patch. handleCaptureSnapshotReply is where the replicatedToAllIndex is set in the behavior. However, In the output, notice that RecoveryCompleted comes after CaptureSnapshotReply. On RecoveryCompleted, we initialize the behavior which creates a new Follower with replicatedToAllIndex initialized to -1. Hence the failure. Recovery is done async and, in local runs, I see RecoveryCompleted occur at various points but if it occurs after CaptureSnapshotReply completes it will fail. We need to wait for recovery to be completed. I can take care of this in my other patch.

Patch-set: 1