Update patch set 2
Patch Set 2:

So the MessageTracker is enabled on Shard startup (i.e. recovery complete) if Shard trace logging is enabled. So you have to assume the controller will become the leader prior to starting it. Of course with the dummy followers this will be the case but not guaranteed otherwise. Also you have to enable trace logging for Shard which will be a lot of output.

How about if we use a separate logging category to enable/disable the MessageTracker?  So in MessageTracker#received:

  private Logger log = LoggerFactory.getLogger(MessageTracker.class);

  public Context received(Object message){
        if(enabled) {
            if(!log.isTraceEnabled()) {
                enabled = false;
                expectedMessageWatch = null;
                return NO_OP_CONTEXT;
            }
       } else if(log.isTraceEnabled()) {
           enabled = true;
           expectedMessageWatch = Stopwatch.createStarted();
       } else {
           return NO_OP_CONTEXT;
       }
      ...
  }

Patch-set: 2
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0