Update patch set 6
Patch Set 6:

My only concern is that the scope for slf4j simple was changed to compile. It really should be test as slf4j simple is not used in production. Also the simpleloggper.props file was put into src/main/resources which I believe will cause it to be included in the production bundle artifact. It probably wouldn't cause an issue but it's not clean.

Also the TestDriver classes get shipped with the production bundle artifact. Personally I'd prefer to move the TestDriver et al under src/test/java or move it to its own sub-project that isn't shipped as a bundle.

Patch-set: 6