Update patch set 2
Patch Set 2:

This is good to go but need to get past the unrelated restconf test failures to get a verification +1. Maybe a rebase will fix it.

I also want to include this new property in the cfg file and add it to the unit tests. I can do that in another patch.

Patch-set: 2
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0