Bug 2769: Fix issues with ApplyLogEntries
ApplyLogEntries defines toIndex as an int which is inconsistent with
everything else that stores journal indexes as long. In addition this
would cause an insiduous bug if an index was greater than max-int.

To maintain backwards compatibility, we can't change the type in
ApplyLogEntries so I created a new class, ApplyJournalEntries, that
defines toIndex as a long and deprecated ApplyLogEntries.

Also, the serialVersionUID for ApplyLogEntries was previously added in
Lithium however this would cause de-serialization failures when
recovering a pre-Lithium journal from persistence. So I removed the
serialVersionUID.

Change-Id: I7c5fe3a2ef8de292224a1a278abe53fc774a79d8
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>