Update patch set 1
Patch Set 1:

My question about why it's necessary to add a dependency is more of a rhetorical question - I wouldn't expect anyone on this patch to know the answer :) Just seems odd to me. My apologies for inadvertently breaking the integration deploy and causing you guys extra work and thanks for tracking this down. I had no idea I had to add a dependency to the features pom.

Patch-set: 1