Introduce a mechanism for a Follower to signal it's sync up status
This patch introduces a new message FollowerInitialSyncUpStatus. The
purpose of the message is to inform the RaftActor or a sub-class thereof
whether the Follower's commitIndex is now atleast at the same level
as the Leader was when it first sent a heartbeat to the Follower.

This will hopefully be useful in a rolling upgrade scenario where
we may have just brought up a new node and need to wait for it to be
synced with the current leader before we bring down and upgrade another
node in the cluster.

Change-Id: If19e4d98c2be708fc6a35823ca92cfac7ca87394
Signed-off-by: Moiz Raja <moraja@cisco.com>