Update patch set 1
Patch Set 1:

I am assuming that you will have another commit that will have changes on Shard and ShardManager that updates MBean to say if a node is synced up or not.

While this implementation looks good, I am wondering if instead of sending a message (FollowerInitialSyncUpStatus) to self (handled by subclass Shard), would a callback such as onSynced() (implemented in subclass) be simpler?

Patch-set: 1