Update patch set 1
Patch Set 1:

Abhishek, there will be a separate patch which put's the value in the ShardStats bean. 

The reason for the message instead of callback is two fold (a) This is not necessarily urgent (b) Follower is not an actor.

Patch-set: 1
Reviewer: Gerrit User 795 <795@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0