Refactor LegacyTransactionConntextImpl
We have a compat package so the LegacyTransactionConntextImpl really
belongs there to keep it separate from the current classes. Also Legacy
really isn't a good name as it's too generic. The
LegacyTransactionConntextImpl is really specific to  pre-Lithium
compatibility so I renamed it to PreLithiumTransactionContextImpl to
make it clear.

I also refactored the helium compatibility tests in TransactionProxyTest
into a new PreLithiumTransactionProxyTest class to keep them separate.
This will make it easier to remove PreLithiumTransactionContextImpl and
its test code when it's no longer needed. Since it needs the same setup
as TransactionProxyTest and some of the utility methods, I moved all the
setup and utility methods to an AbstractTransactionProxyTest to avoid
duplication. This also makes TransactionProxyTest smaller as it now only
contains the test methods.

Change-Id: I9a81c7a1a07fac9098d497ff301c5d1909094c1b
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>