Update patch set 5
Patch Set 5: -Code-Review

I was thinking of something in ShardManager MBean that says current node is synced up with the leader. This it does by checking if all the Shards it knows about have synced up or not.
This will be useful when the client only wants to know the status of node (synced up or not) and doesn't know or care what shards exist on it.

Patch-set: 5
Reviewer: Gerrit User 807 <807@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0