Refactor ShardTest
Refactored the helium compatibility tests in ShardTest into a new
PreLithiumShardTest class to keep them separate. This will make it
easier to remove the Helium test code when it's no longer needed.
Since it needs the same setup as ShardTest and some of the utility
methods, I moved all the setup and utility methods to an
AbstractShardTest to avoid duplication. This also makes ShardTest
smaller and easier to follow.

I also moved the ShardTransactionHeliumBackwardsCompatibilityTest to
the compat package.

Change-Id: Ie46441adbd0be4c4a655f5db640519705a6ced29
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>