Modify FindPrimary to check for leader
Write-only transaactions now bypass sending the CreateTransaction
message to the shard however we still need to check if the shard has a leader.
So I modified the ShardManager to check for this on FindPrimaryShard
message utilizing the role changed events recently added. On FindPrimaryShard,
if the shard is not ready, ie its role is Candidate or not initialized, the sender
is recorded and a timeout schedule is started. If the shard becomes
ready, the success message is sent to the sender. If the timer expires, an error message
is sent to the sender, either ActorNotInitilized or
NoShardLeaderException.

Change-Id: I22bc661314b25812024b7d274afb390768408b0b
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>