Update patch set 9
Patch Set 9:

I was asked by Tony if we could push a later patch with JUnit tests since this patch is blocking other downstream patches.

In general I disagree with this approach as it leaves the new code uncovered and risks regressions down the line especially if JUnits are not added. In the future I personally will not be so willing to accept a patch without unit tests.

That said, since this patch is blocking other patches related to deprecating APIs (per Tony) I do agree that we can push this patch forward with the following stipulations:

1) Marian adds comments to this gerrit that JUnits will be following immediately with another gerrit patch AND the same is noted in the BUG.
2) The BUG is not closed until all JUnits are pushed and merged.

IF we do the above and if we have addressed all other comments then we can merge this in.

Patch-set: 9
Label: Verified=0