BUG-2673: make CDS implement DOMDataTreeChangeListener
Adds the ability to transport tree changes efficiently across the
network and deliver them to clients. Only bare minimum deltas are
transmitted. This implementation does not completely reconstruct the
before and after images required for clients -- that will be addressed
by a separate follow-up patch. The entry place where that patch will
resume its work is marked with a FIXME.

Change-Id: Ifb91d08857684fb160fd923bc25c294d2fca4bc3
Signed-off-by: Robert Varga <rovarga@cisco.com>