Update patch set 1
Patch Set 1: Code-Review-1

I dont think going the way of reconstructing the hello message is the best way to implement this workaround. I would remove the namespace check and then in NetconfClientSessionNegotiator add a second XPath without namespaces that would be used as fallback when the XPath with namespaces fails to find the session-id.

Patch-set: 1
Label: Code-Review=-1