Update patch set 21
Patch Set 21:

Just talked to Ed and he informed me that this patch is a blocker for the CDS switchover as BGP needs it. I didn't know the new change listener API was actually gong to be used in Li - I thought it was essentially a POC at this point.

The code looks good at this point. I can't system test it though unless I take some time to write a  test app that uses it. I'll feel more comfortable with it when there are unit tests which I can work on next week. In the mean time, Robert - have you been system testing this with BGP? If so and if you're comfortable with the results, we can merge it although it would be good if Moiz can review as well before doing so.

Patch-set: 21
Label: Verified=0