Update patch set 21
Patch Set 21:

I hear ya Moiz. We can get +1's on this patch and essentially call it complete but not merge it yet until unit tests are done. I can do that in a separate patch. Or if Robert wants to write the tests, that's fine. 

How does that sound?

Patch-set: 21