DO NOT SUBMIT : ClusteringService via Notifications
Clustering Service which notifies the subscribers of the role changes 
via Notifications

A new actor ShardRoleChangeListener is created, by the ShardManager,
which listens to the RoleChangeNotifier for role change notifications.

On receipt it calls the rpc exposed by the clustering service.

ClusteringService rpc raises notifications to its subscribers.

Tests are to be added.

Change-Id: I7e056322b549b78459728ccbe5ddde247078bccd
Signed-off-by: Kamal Rameshan <kramesha@cisco.com>