Update patch set 6
Patch Set 6:

Thanks, Yes, I rebased patch yesterday. I am working on exposing of EventSourceRegistry as a service and it could influence my decision about interface. EventSourceRegistry is part of api and it should be interface, I agree with you. But EventSourceRegistration extends another abstract class and it slightly complicates solution. I want to solve it today and push it.

Patch-set: 6