Update patch set 1
Patch Set 1: Code-Review-1

Is there a reason that some HTTP methods, e.g. TRACE and OPTIONS, are not covered by the list of http-method elements? If not, and the intention is to require authn for all methods, then it would be better to remove all of the http-method elements. The default behavior of the security-constraint element is to include all HTTP methods; it only applies to particular methods if they are explicitly enumerated.

Patch-set: 1
Label: Code-Review=-1